requirements.txt
omnidimension
python-dotenv
requests